:recycle: Refactorizacion del codigo para las distintas vistas
