+++ client/package.json
@@ -14,6 +14,6 @@
     "vite": "^4.4.5"
   },
   "dependencies": {
-    "file-saver": "^2.0.5"
+    "sweetalert2": "^11.7.31"
   }
 }
