+++ client/styles/components/hero-banner.scss
@@ -1,5 +1,3 @@
-
-
 .hero-banner{
     background: linear-gradient(0deg, rgba(0, 0, 0, 0.35) 0%, rgba(0, 0, 0, 0.35) 100%), url('../../assets/fondo-banner-1.jpg') no-repeat center/cover, lightgray 50%;
     gap: 32px;
