+++ client/styles/elements.scss
@@ -50,6 +50,7 @@ ion-icon{
 /* Header */
 
 header {
+    position: relative;
     width: 100%;
     border-bottom: 1px solid var(--shark-skin);
     position: sticky;
