+++ client/views/about-us.html
@@ -94,9 +94,9 @@
         </ul>
       </div>
     </section>
-  </header>
 
-  <!-- Menu mobile -->
+
+    <!-- Menu mobile -->
   <section id="menu-mobile" class="mobile-menu hide">
     <div class="mobile-menu__container">
       <div class="mobile-menu__brand space-between">
@@ -152,13 +152,13 @@
       <div class="mobile-menu__navigation">
         <div class="mobile-menu__pages">
           <span>
-            <img src="/assets/icons/icon-gift.svg" alt="">
+            <img src="../assets/icons/icon-gift.svg" alt="">
           </span>
           <a class="fs-xs" href="">Blog</a>
         </div>
         <div class="mobile-menu__pages">
           <span>
-            <img src="/assets/icons/icon-like.svg" alt="">
+            <img src="../assets/icons/icon-like.svg" alt="">
           </span>
           <a class="fs-xs" href="">Blog</a>
         </div>
@@ -170,6 +170,7 @@
       </div>
     </div>
   </section>
+  </header>
 
   <main class=" ">
     <div class="about__container-title main__container">
