+++ client/views/add-products.html
@@ -12,6 +12,7 @@
   <link
     href="https://fonts.googleapis.com/css2?family=Montserrat:wght@400;500;600;700&family=Poppins:wght@400;500;600;700&display=swap"
     rel="stylesheet">
+    <script src="https://cdn.jsdelivr.net/npm/sweetalert2@11"></script>
 </head>
 
 <body>
@@ -314,9 +315,9 @@
   </footer>
 
   <script type="module" src="https://unpkg.com/ionicons@7.1.0/dist/ionicons/ionicons.esm.js"></script>
-  <script src="https://formspree.io/js/formbutton-v1.min.js" defer></script>
-  <!-- <script type="module" src="../public/main.js"></script> -->
+  <script type="module" src="../public/main.js"></script>
   <script type="module" src="../scripts/addProduct.js"></script>
+
 </body>
 
 </html>
