+++ client/views/contact-us.html
@@ -3,7 +3,7 @@
 
 <head>
   <meta charset="UTF-8" />
-  <link rel="icon" type="image/svg+xml" href="/assets/icons/icon-logo.svg" />
+  <link rel="icon" type="image/svg+xml" href="../assets/icons/icon-logo.svg" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <title>Contact Us</title>
   <link rel="stylesheet" href="../public/style.css">
@@ -94,9 +94,9 @@
         </ul>
       </div>
     </section>
-  </header>
 
-  <!-- Menu mobile -->
+
+    <!-- Menu mobile -->
   <section id="menu-mobile" class="mobile-menu hide">
     <div class="mobile-menu__container">
       <div class="mobile-menu__brand space-between">
@@ -152,13 +152,13 @@
       <div class="mobile-menu__navigation">
         <div class="mobile-menu__pages">
           <span>
-            <img src="/assets/icons/icon-gift.svg" alt="">
+            <img src="../assets/icons/icon-gift.svg" alt="">
           </span>
           <a class="fs-xs" href="">Blog</a>
         </div>
         <div class="mobile-menu__pages">
           <span>
-            <img src="/assets/icons/icon-like.svg" alt="">
+            <img src="../assets/icons/icon-like.svg" alt="">
           </span>
           <a class="fs-xs" href="">Blog</a>
         </div>
@@ -170,7 +170,7 @@
       </div>
     </div>
   </section>
-
+  </header>
   <main class="main__container">
     <form id="contact-form" class="left formulario-contacto" action="https://formspree.io/f/xdorpkld" method="POST">
       <div class="left formulario-contacto__descripcion">
@@ -288,7 +288,6 @@
 
     </div>
   </main>
-  
   <footer>
     <div class="footer__container-navigation">
       <div class="footer__header ft-14">
@@ -337,10 +336,8 @@
       <p class="ft-medium fs-xs">Made With Love By Dynamite Code All Right Reserved </p>
     </div>
   </footer>
-
   <script type="module" src="https://unpkg.com/ionicons@7.1.0/dist/ionicons/ionicons.esm.js"></script>
   <script src="https://formspree.io/js/formbutton-v1.min.js" defer></script>
   <script type="module" src="../public/main.js"></script>
 </body>
-
 </html>
