+++ client/views/home.html
@@ -94,9 +94,9 @@
         </ul>
       </div>
     </section>
-  </header>
 
-  <!-- Menu mobile -->
+
+    <!-- Menu mobile -->
   <section id="menu-mobile" class="mobile-menu hide">
     <div class="mobile-menu__container">
       <div class="mobile-menu__brand space-between">
@@ -170,6 +170,7 @@
       </div>
     </div>
   </section>
+  </header>
 
   <main class="main__container">
     <section class="promociones">
@@ -244,6 +245,7 @@
 
   <script type="module" src="https://unpkg.com/ionicons@7.1.0/dist/ionicons/ionicons.esm.js"></script>
   <script type="module" src="../public/main.js"></script>
+  <script type="module" src="../scripts/home.js"></script>
 </body>
 
 </html>
